fix: modals on admin page and css minor changes
