+++ src/components/features/ModalDeleteCategory/Styles.js
@@ -5,6 +5,7 @@ export const Container = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  gap: 3em;
 
   width: 100%;
   max-width: 1000px;
@@ -16,6 +17,7 @@ export const Message = styled.h1`
   font-family: ${(props) => props.theme.fonts.montserrat};
   text-align: center;
   margin: 1em;
+  color: white;
 `;
 
 export const Delete = styled.button`
@@ -23,10 +25,10 @@ export const Delete = styled.button`
   height: 3em;
 
   border-radius: 8px;
-  background-color: red;
+  background-color: white;
 
   font-size: 1rem;
   font-family: ${(props) => props.theme.fonts.montserrat};
-  font-weight: 500;
+  font-weight: 700;
   text-align: center;
 `;
