+++ src/pages/Catalog/Catalog.jsx
@@ -1,13 +1,6 @@
-import { useState } from 'react';
-
-import { CloseOutlined } from '@ant-design/icons';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 
-import {
-  ModalDeleteCategory,
-  ModalDeleteProduct,
-} from '../../components/features';
 import { useGetCategories } from '../../hooks/query/categories';
 import {
   Page,
@@ -25,27 +18,11 @@ import {
   Product,
   ProductImage,
   ProductName,
-  ModalStyle,
 } from './Styles';
 import buildGetCategoriesErrorMessage from './utils';
 
 export default function Catalog() {
   const navigate = useNavigate();
-  const [modalDeleteProduct, setModalDeleteProduct] = useState(false);
-  const [modalDeleteCategory, setModalDeleteCategory] = useState(false);
-  const [productId, setProductId] = useState('');
-  const [categoryId, setCategoryId] = useState('');
-
-  const openModalDeleteProduct = (_id) => {
-    setProductId(_id);
-    setModalDeleteProduct(true);
-  };
-  const openModalDeleteCategory = (_id) => {
-    setCategoryId(_id);
-    setModalDeleteCategory(true);
-  };
-  const closeModalDeleteProduct = () => setModalDeleteProduct(false);
-  const closeModalDeleteCategory = () => setModalDeleteCategory(false);
 
   const { data: categories, isLoading } = useGetCategories({
     onError: (err) => {
@@ -83,13 +60,7 @@ export default function Catalog() {
             <ButtonRow>
               {categories?.map((category) => (
                 <Anchor key={category.name} href={`#${category.name}`}>
-                  <Button>
-                    {category.name}
-                    <CloseOutlined
-                      style={{ fontSize: '1rem' }}
-                      onClick={() => openModalDeleteCategory(category._id)}
-                    />
-                  </Button>
+                  <Button>{category.name}</Button>
                 </Anchor>
               ))}
             </ButtonRow>
@@ -99,19 +70,12 @@ export default function Catalog() {
                 <CategoryName>{category.name}</CategoryName>
                 <ProductRow>
                   {category?.products?.map((product) => (
-                    <Product key={product.name}>
-                      <CloseOutlined
-                        onClick={() => openModalDeleteProduct(product._id)}
-                      />
-                      <ProductImage
-                        src={product.pictures[0].url}
-                        onClick={() => navigate(`/produto/${product._id}`)}
-                      />
-                      <ProductName
-                        onClick={() => navigate(`/produto/${product._id}`)}
-                      >
-                        {product.name}
-                      </ProductName>
+                    <Product
+                      key={product.name}
+                      onClick={() => navigate(`/produto/${product._id}`)}
+                    >
+                      <ProductImage src={product.pictures[0].url} />
+                      <ProductName>{product.name}</ProductName>
                     </Product>
                   ))}
                 </ProductRow>
@@ -120,31 +84,6 @@ export default function Catalog() {
           </>
         )}
       </Container>
-      <ModalStyle
-        open={modalDeleteProduct}
-        onCancel={closeModalDeleteProduct}
-        footer={null}
-        width={1000}
-        closeIcon={<CloseOutlined />}
-        destroyOnClose
-        centered
-      >
-        <ModalDeleteProduct _id={productId} close={closeModalDeleteProduct} />
-      </ModalStyle>
-      <ModalStyle
-        open={modalDeleteCategory}
-        onCancel={closeModalDeleteCategory}
-        footer={null}
-        width={1000}
-        closeIcon={<CloseOutlined />}
-        destroyOnClose
-        centered
-      >
-        <ModalDeleteCategory
-          _id={categoryId}
-          close={closeModalDeleteCategory}
-        />
-      </ModalStyle>
     </Page>
   );
 }
