+++ src/pages/ListProduct/ListProduct.jsx
@@ -7,7 +7,11 @@ import { useMediaQuery } from 'react-responsive';
 import { toast } from 'react-toastify';
 
 import { Select } from '../../components/common';
-import { ModalEditProduct } from '../../components/features';
+import {
+  ModalDeleteCategory,
+  ModalDeleteProduct,
+  ModalEditProduct,
+} from '../../components/features';
 import { useGetCategories } from '../../hooks/query/categories';
 import { useSearchProductByName } from '../../hooks/query/products';
 import useDebounce from '../../hooks/useDebounce';
@@ -26,6 +30,7 @@ import {
   EditButton,
   SearchIconButton,
   ModalStyle,
+  DeleteButton,
 } from './Styles';
 import {
   buildGetProductsErrorMessage,
@@ -35,6 +40,10 @@ import {
 export default function ListProduct() {
   const isSmallScreen = useMediaQuery({ maxWidth: 700 });
   const [selectedCategory, setSelectedCategory] = useState({});
+  const [modalDeleteProduct, setModalDeleteProduct] = useState(false);
+  const [productId, setProductId] = useState('');
+  const [modalDeleteCategory, setModalDeleteCategory] = useState(false);
+  const [categoryId, setCategoryId] = useState('');
   const [modalEditProduct, setModalEditProduct] = useState(false);
   const [selectedProduct, setSelectedProduct] = useState({});
   const [name, setName] = useState('');
@@ -58,11 +67,23 @@ export default function ListProduct() {
     },
   });
 
+  const openModalDeleteProduct = (_id) => {
+    setProductId(_id);
+    setModalDeleteProduct(true);
+  };
+
+  const openModalDeleteCategory = (_id) => {
+    setCategoryId(_id);
+    setModalDeleteCategory(true);
+  };
+
   const openModalEditProduct = (product) => {
     setSelectedProduct(product);
     setModalEditProduct(true);
   };
   const closeModalEditProduct = () => setModalEditProduct(false);
+  const closeModalDeleteProduct = () => setModalDeleteProduct(false);
+  const closeModalDeleteCategory = () => setModalDeleteCategory(false);
 
   return (
     <Container>
@@ -94,6 +115,11 @@ export default function ListProduct() {
         {products?.map((product) => (
           <Row key={product._id}>
             <Text>{product.name}</Text>
+            <DeleteButton>
+              <CloseOutlined
+                onClick={() => openModalDeleteCategory(product.category._id)}
+              />
+            </DeleteButton>
             <Text>{product.category.name}</Text>
 
             {isSmallScreen ? (
@@ -101,12 +127,19 @@ export default function ListProduct() {
                 <TbPencil size={30} />
               </StyledLink>
             ) : (
-              <EditButton>
-                <TbPencil
-                  onClick={() => openModalEditProduct(product)}
-                  size={30}
-                />
-              </EditButton>
+              <>
+                <DeleteButton>
+                  <CloseOutlined
+                    onClick={() => openModalDeleteProduct(product._id)}
+                  />
+                </DeleteButton>
+                <EditButton>
+                  <TbPencil
+                    onClick={() => openModalEditProduct(product)}
+                    size={30}
+                  />
+                </EditButton>
+              </>
             )}
           </Row>
         ))}
@@ -133,6 +166,31 @@ export default function ListProduct() {
         <ModalEditProduct
           product={selectedProduct}
           close={closeModalEditProduct}
+        />
+      </ModalStyle>
+      <ModalStyle
+        open={modalDeleteProduct}
+        onCancel={closeModalDeleteProduct}
+        footer={null}
+        width={1000}
+        closeIcon={<CloseOutlined />}
+        destroyOnClose
+        centered
+      >
+        <ModalDeleteProduct _id={productId} close={closeModalDeleteProduct} />
+      </ModalStyle>
+      <ModalStyle
+        open={modalDeleteCategory}
+        onCancel={closeModalDeleteCategory}
+        footer={null}
+        width={1000}
+        closeIcon={<CloseOutlined />}
+        destroyOnClose
+        centered
+      >
+        <ModalDeleteCategory
+          _id={categoryId}
+          close={closeModalDeleteCategory}
         />
       </ModalStyle>
     </Container>
